Fix remote command checking and remove duplicate TRAMP utility function
This commit fixes a bug where remote command checking always searched locally,
and removes the duplicate remote-specific function that was working around this
bug.

Bug Fix:
  - Fix core/core-utils.el core-utils-check-command-in-path to pass is-remote
    flag to executable-find, enabling proper remote host command detection via
    TRAMP

Code Cleanup:
  - Remove duplicate core-utils-check-command-in-path-remote-host function from
    features/tramp/tramp-utils.el (16 lines)
  - Simplify features/eglot/eglot-config.el to use unified command checking
    function for both local and remote files

The core function now correctly searches on the remote host when
default-directory is remote, eliminating the need for a separate remote-only
version.

Signed-off-by: Shane Bradley <68629+sbradley7777@users.noreply.github.com>